requirements.txt
# Core dependencies
pandas>=2.0.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0


# Optional for enhanced features
python-dotenv>=1.0.0
structlog>=23.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Streamlit dashboard
streamlit>=1.28.0
plotly>=5.15.0